Split ConfigDispatcher into ConfigDispatcher and ConfigDispatcherFileWatcher
Separate the processing of file-system events from our ConfigDispatcher
tasks, to simplify tests.
